requirements.txt
flask
pytesseract
pdf2image
Pillow
PyMuPDF
gunicorn
